Implement a state machine top-level
Not sure if I want to put it in game.pump, but it has to be part of game, as it's the game logic itself that drives what state we're in.
The state is evaluated using a memoization (since we tend to not assign functions). In addition, we'd need this to prevent a table lookup (which is boring).

Right now this is the best design I can come up with. It's completely general and still allows the game state/stack to be investigated.

Regarding that: I need to make it a stack.