Implement tile drawing
Tile drawing is currently VERY unoptimized and so, VERY slow.
We'll need to draw all tiles to some texture and then use that texture, which
would appear to be the fastest way to do things.

Right now we just call drawing functions in a loop, these drawing functions are
not meant to be called in a loop.