Implement perspective change handling
When the screen size changes, the game will need to be updated,
for instance, we need to add a "perspective" matrix that ensures
that all the items remain in their consistent sizes.

Now, the perspective matrix can be used for even more, for instance,
to be able to view more of the field, which can be done by a scaling
matrix.