Add the architecture to graph
I think I've stumbled upon a quite supreme architecture. By leveraging HSMs
we can make every state a dumb pipe that uses a pure and impure library.

This offloads most "real problems" to libraries while keeping each state
"visible" as in - we can EASILY see what happens in each state, since these
are forced to be pipelines.

Of course, you could just create a huge library func with whatever you
want to do, but this is clearly against the spirit in this scheme.