Make should-exit a compound variable
Previously we simply overwrote should-exit, which made the result from
glfwWindowShouldClose useless. Now we combine the values.

The nice thing about this is that it's actually testable! We run the
cores function once with `ae.should-exit?` true and we look to see if
it tried to pop the fsm system.