Actually test core-pure
This tests if should-exit remains through throughout the computation.
This is important because regressions have happened and it is
annoying when it happens. This test will render such regressions
impossible.

I'll need a better language for this though, something like matching a
substructure. I like the fixture idea. This makes the transformation
easy to handle, no magic is required.

How then do we handle the IO classes? Give them a fixture?
Not liking that idea. We should separate concerns in such a manner
that they can actually be tested using their real implementations.

That would be fantastic. I just need to forgo a "setup based fixture"
or something. I really don't want that.