Migrate design into "utils"
The new design is:

        The Mediator (states, coarse logic)
        The Utilities (fine-grained logic)

We now have a path called "pgame/utils" that includes every single file
from the utilities directory. I think I'll need to add ignores for
compiled directories though. The cool thing about this is:

        We can add a new file, use any language we wish, and it'll be
        available everywhere.

All that other parts of the program needs are `(require pgame/utils)`.
Ideally, this would also be automated, however, it currently is not.

What this design allows is for us to create a massive library with ease,
and use all functions without wasting time importing stuff.